move changes to semanticplaceEe env
